XIVY-11947 fix NotificationSubscription having moved in core
